requirements.txt
telethon
python-dotenv
anthropic
supabase
boto3
python-telegram-bot[job-queue]
